# api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mini RBAC Go API",
    "version": "2.0.0",
    "description": "Role-Based Access Control API for managing roles, groups, permissions, and workspace hierarchies",
    "contact": {
      "name": "Mini RBAC Go",
      "url": "https://github.com/redhat/mini-rbac-go"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8085/api/rbac/v2",
      "description": "Local development server"
    }
  ],
...
  "paths": {
    "/roles": {
      "get": {
        "tags": ["Roles"],
        "summary": "List all roles",
        "description": "Returns a paginated list of roles for the tenant",
        "operationId": "listRoles",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Roles"],
        "summary": "Create a new role",
        "description": "Creates a new custom role with specified permissions",
        "operationId": "createRole",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Role created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/roles/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/RoleId"
        }
      ],
      "get": {
        "tags": ["Roles"],
        "summary": "Get role by ID",
        "description": "Returns a single role with its permissions",
        "operationId": "getRole",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": ["Roles"],
        "summary": "Update role",
        "description": "Updates an existing role's name, description, or permissions",
        "operationId": "updateRole",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": ["Roles"],
        "summary": "Delete role",
        "description": "Deletes a role and all its associations",
        "operationId": "deleteRole",
        "responses": {
          "204": {
            "description": "Role deleted successfully"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Role unique identifier"
          },
          "name": {
            "type": "string",
            "description": "Role name"
          },
          "description": {
            "type": "string",
            "description": "Role description"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          },
          "last_modified": {
            "type": "string",
            "format": "date-time",
            "description": "Last modification timestamp"
          }
        }
      },
      "Permission": {
        "type": "object",
        "properties": {
          "application": {
            "type": "string",
            "description": "Application name",
            "example": "inventory"
          },
          "resource_type": {
            "type": "string",
            "description": "Resource type",
            "example": "hosts"
          },
          "permission": {
            "type": "string",
            "description": "Permission/verb",
            "example": "read"
          }
        },
        "required": ["application", "resource_type", "permission"]
      },
      "CreateRoleRequest": {
        "type": "object",
        "required": ["name", "description", "permissions"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Role name",
            "example": "Inventory Manager"
          },
          "description": {
            "type": "string",
            "description": "Role description",
            "example": "Manage inventory resources"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            },
            "minItems": 1
          }
        }
      },
      "UpdateRoleRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Updated role name"
          },
          "description": {
            "type": "string",
            "description": "Updated role description"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          }
        }
      },
...
      "RoleListResponse": {
        "type": "object",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          }
        }
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "description": "Number of items in current response"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Error title"
          },
          "detail": {
            "type": "string",
            "description": "Detailed error message"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code"
          }
        }
      }
    },
    "parameters": {
      "RoleId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Role UUID",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}